Web Build 2 Added For Real This Time
Had to comment out a part of the git ignore in order commit builds folders. I am not sure if this is the best way to do this, but I like using github pages to show people stuff and this is the only way I know how. Will look into different options later