Looping Sound Works, Started on 3.0, and Changed Blue to have Spawn Points
Beat objects now have the option to specify if the sound is one beat or not. If it is not it will just loop the one clip over and over instead of playing every time. 3.0 is starting to be made, blue platforms now operate based on spawn points. You can set how many will stay on the screen at one time. It will move the one that has existed the longest back to the orginial pos when the beat plays